+++ sccfm_core/services/__init__.py
@@ -9,6 +9,7 @@ from sccfm_core.services.inventory import (
     AsaUpgradeVersionService,
     AsaUserPasswordService,
     InventoryService,
+    ShunEntrySpec,
 )
 from sccfm_core.services.object_management import (
     NetworkGroupListResponse,
@@ -24,8 +25,8 @@ __all__ = [
     "AsaBootRegistryService",
     "AsaCommandLineService",
     "AsaDiskFileService",
-    "AsaUpgradeService",
     "AsaShunService",
+    "AsaUpgradeService",
     "AsaUpgradeVersionService",
     "AsaUserPasswordService",
     "HealthService",
@@ -37,4 +38,5 @@ __all__ = [
     "NetworkObjectListResponse",
     "NetworkObjectResponse",
     "NetworkObjectService",
+    "ShunEntrySpec",
 ]